Duplicate handling

+++ src/store/reducers/ormReducer/index.js
@@ -115,34 +115,61 @@ export default function ormReducer (state = {}, action) {
       break
 
     case CREATE_TOPIC_PENDING:
-      topic = Topic.create({
-        id: meta.tempId,
-        name: meta.topicName,
-        postsTotal: 0,
-        followersTotal: 1
-      })
+      // Current backend is a "find or create", so we need to check if we
+      // already have it on the front end to avoid duplicates
+      // TODO: review when topics are a group
+      const topicQuery = Topic.all().filter({ name: meta.topicName })
+      topic = topicQuery.count() === 0
+        ? Topic.create({
+          id: meta.tempId,
+          name: meta.topicName,
+          postsTotal: 0,
+          followersTotal: 1
+        })
+        : topicQuery.first()
+
       community = Community.withId(meta.communityId)
-      CommunityTopic.create({
-        id: meta.communityTopicTempId,
-        community,
-        followersTotal: 1,
-        isSubscribed: true,
-        postsTotal: 0,
-        topic
-      })
+      // Also check for an existing CommunityTopic for the same reason
+      const ctQueryPending = CommunityTopic.all().filter({ community, topic })
+      if (ctQueryPending.count() === 0) {
+        CommunityTopic.create({
+          id: meta.communityTopicTempId,
+          community,
+          followersTotal: 1,
+          isSubscribed: true,
+          postsTotal: 0,
+          topic
+        })
+      }
       break
 
     case CREATE_TOPIC:
-      Topic.withId(meta.tempId).delete()
-      topic = Topic.create(payload.data.createTopic)
+      // Scenarios we need to handle here:
+      //  - the Topic may or may not exist in the client (temp or actual)
+      //  - the CommunityTopic may or may not exist in the client (temp or actual)
+      // TODO: review when topics are a group
+      const { createTopic } = payload.data
+      const communityTopic = createTopic.communityTopics.items[0]
       community = Community.withId(meta.communityId)
-      CommunityTopic.withId(meta.communityTopicTempId).delete()
-      CommunityTopic.create({
-        ...payload.data.createTopic.communityTopics.items[0],
-        community,
-        topic,
-        isSubscribed: true
-      })
+      topic = Topic.hasId(createTopic.id)
+        ? Topic.withId(createTopic.id)
+        : Topic.create(createTopic)
+
+      if (Topic.hasId(meta.tempId)) {
+        Topic.withId(meta.tempId).delete()
+      }
+      if (CommunityTopic.hasId(meta.communityTopicTempId)) {
+        CommunityTopic.withId(meta.communityTopicTempId).delete()
+      }
+
+      if (!CommunityTopic.hasId(communityTopic.id)) {
+        CommunityTopic.create({
+          ...communityTopic,
+          community,
+          topic,
+          isSubscribed: true
+        })
+      }
       break
 
     case FETCH_MESSAGES_PENDING:
